# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  docs.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
  docs.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
...
host: router.yeying.pub
info:
  contact: {}
  description: Router public/admin API (JWT/UCAN) under /api/v1.
  title: Router API
  version: v1
paths:
...
  /api/v1/admin/dashboard:
    get:
      parameters:
      - description: today|last_7_days|last_30_days|this_month|last_month|this_year|last_year|last_12_months|all_time
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docs.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get admin dashboard aggregate data
      tags:
      - admin
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"